Move artifact manifest code to artifact_manifest.py
