+++ wandb/sdk/interface/artifacts/__init__.py
@@ -1,3 +1,4 @@
-from .artifacts import Artifact, ArtifactEntry, ArtifactManifest
+from .artifact import Artifact
+from .artifact_manifest import ArtifactEntry, ArtifactManifest
 from .cache import ArtifactsCache, get_artifacts_cache
 from .storage import StorageHandler, StorageLayout, StoragePolicy
